refactor(character): split CharacterSheetSections into section/ package
Each sheet section composable lives in its own file under the section/
subpackage. CombatSection.kt groups CombatRow and WalkSpeedRow together.
