+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/character/presentation/component/CharacterDetailScreen.kt
@@ -31,6 +31,12 @@ import com.cyrillrx.rpg.character.presentation.CharacterEditState
 import com.cyrillrx.rpg.character.presentation.CharacterEditState.Loaded.EditingField
 import com.cyrillrx.rpg.character.presentation.CoercedValue
 import com.cyrillrx.rpg.character.presentation.component.dialog.CharacterEditDialog
+import com.cyrillrx.rpg.character.presentation.component.section.AbilityGrid
+import com.cyrillrx.rpg.character.presentation.component.section.CombatRow
+import com.cyrillrx.rpg.character.presentation.component.section.LanguagesRow
+import com.cyrillrx.rpg.character.presentation.component.section.SavingThrowsSection
+import com.cyrillrx.rpg.character.presentation.component.section.SheetDivider
+import com.cyrillrx.rpg.character.presentation.component.section.WalkSpeedRow
 import com.cyrillrx.rpg.character.presentation.navigation.CharacterRouter
 import com.cyrillrx.rpg.character.presentation.viewmodel.CharacterEditViewModel
 import com.cyrillrx.rpg.core.presentation.LocalDistanceUnit
