+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/character/presentation/component/section/CombatSection.kt
@@ -1,0 +1,105 @@
+package com.cyrillrx.rpg.character.presentation.component.section
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import com.cyrillrx.rpg.character.presentation.component.StatCell
+import com.cyrillrx.rpg.core.domain.toSignedString
+import com.cyrillrx.rpg.core.presentation.LocalDistanceUnit
+import com.cyrillrx.rpg.core.presentation.component.dnd.toDistanceString
+import com.cyrillrx.rpg.core.presentation.theme.borderAlpha
+import com.cyrillrx.rpg.core.presentation.theme.borderWidth
+import com.cyrillrx.rpg.core.presentation.theme.spacingCommon
+import com.cyrillrx.rpg.core.presentation.theme.spacingMedium
+import com.cyrillrx.rpg.settings.domain.DistanceUnit
+import org.jetbrains.compose.resources.stringResource
+import rpg_companion.composeapp.generated.resources.Res
+import rpg_companion.composeapp.generated.resources.label_ac
+import rpg_companion.composeapp.generated.resources.label_initiative
+import rpg_companion.composeapp.generated.resources.label_max_hp
+import rpg_companion.composeapp.generated.resources.label_walk_speed
+import rpg_companion.composeapp.generated.resources.settings_unit_feet_abbr
+import rpg_companion.composeapp.generated.resources.settings_unit_meters_abbr
+
+@Composable
+internal fun CombatRow(
+    armorClass: Int,
+    initiative: Int,
+    maxHitPoints: Int,
+    onAcTapped: () -> Unit,
+    onMaxHpTapped: () -> Unit,
+) {
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(spacingMedium),
+        modifier = Modifier.fillMaxWidth(),
+    ) {
+        StatCell(
+            label = stringResource(Res.string.label_ac),
+            value = armorClass.toString(),
+            onClick = onAcTapped,
+            modifier = Modifier.weight(1f),
+        )
+        StatCell(
+            label = stringResource(Res.string.label_initiative),
+            value = initiative.toSignedString(),
+            modifier = Modifier.weight(1f),
+        )
+        StatCell(
+            label = stringResource(Res.string.label_max_hp),
+            value = maxHitPoints.toString(),
+            onClick = onMaxHpTapped,
+            modifier = Modifier.weight(1f),
+        )
+    }
+}
+
+@Composable
+internal fun WalkSpeedRow(
+    walkSpeed: Int,
+    onTap: () -> Unit,
+) {
+    Card(
+        onClick = onTap,
+        border = BorderStroke(borderWidth, MaterialTheme.colorScheme.outline.copy(alpha = borderAlpha)),
+        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
+        modifier = Modifier.fillMaxWidth(),
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = spacingCommon, vertical = spacingMedium),
+        ) {
+            Text(
+                text = stringResource(
+                    Res.string.label_walk_speed,
+                    stringResource(
+                        when (LocalDistanceUnit.current) {
+                            DistanceUnit.FEET -> Res.string.settings_unit_feet_abbr
+                            DistanceUnit.METERS -> Res.string.settings_unit_meters_abbr
+                        },
+                    ),
+                ),
+                style = MaterialTheme.typography.labelSmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                modifier = Modifier.weight(0.5f),
+            )
+            Text(
+                text = walkSpeed.toDistanceString(LocalDistanceUnit.current),
+                style = MaterialTheme.typography.bodyMedium,
+                fontWeight = FontWeight.SemiBold,
+                modifier = Modifier.weight(0.5f),
+            )
+        }
+    }
+}
